moved structs into a separate file
